requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
jwt==1.3.1
pycparser==2.21
PyJWT @ git+https://github.com/jpadilla/pyjwt.git@f86b8b6ce670e40f1ef037b70ac6b4c682e8ac6f
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
